Moved autonomous code to a new file base.py in autonomous/ but kept all the routines in autonomous.py
So all the routines are in one file and all the code behind the routines are in another file
